Rename file to NetworkSelector
