+++ src/components/index.ts
@@ -21,7 +21,7 @@ export { default as Modal } from './Modal';
 export { IButton, LegacyModal } from './LegacyModal';
 export { default as NewAppReleaseModal } from './NewAppReleaseModal';
 export { default as InputField } from './InputField';
-export { default as NetworkSelectDropdown } from './NetworkSelectDropdown';
+export { default as NetworkSelectDropdown } from './NetworkSelector';
 export { default as PrintPaperWalletButton } from './PrintPaperWalletButton';
 export { default as Overlay } from './Overlay';
 export { default as QRCode, QRCodeContainer } from './QRCode';
